+++ lib/widgets/common/map/leaflet/tile_layers.dart
@@ -2,6 +2,20 @@ import 'package:aves/model/device.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_map/flutter_map.dart';
 
+class OpenTopoMapLayer extends StatelessWidget {
+  const OpenTopoMapLayer({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return TileLayer(
+      urlTemplate: 'https://{s}.tile.opentopomap.org/{z}/{x}/{y}.png',
+      subdomains: const ['a', 'b', 'c'],
+      retinaMode: MediaQuery.devicePixelRatioOf(context) > 1,
+      userAgentPackageName: device.userAgent,
+    );
+  }
+}
+
 class OSMHotLayer extends StatelessWidget {
   const OSMHotLayer({super.key});
 
@@ -23,7 +37,6 @@ class StamenWatercolorLayer extends StatelessWidget {
   Widget build(BuildContext context) {
     return TileLayer(
       urlTemplate: 'https://watercolormaps.collection.cooperhewitt.org/tile/watercolor/{z}/{x}/{y}.jpg',
-      subdomains: const ['a', 'b', 'c', 'd'],
       retinaMode: MediaQuery.devicePixelRatioOf(context) > 1,
       userAgentPackageName: device.userAgent,
     );