+++ plugins/aves_map/lib/src/style.dart
@@ -1,10 +1,13 @@
+// OSM raster tile providers: https://wiki.openstreetmap.org/wiki/Raster_tile_providers
+// OSM vector tile providers: https://wiki.openstreetmap.org/wiki/Vector_tiles#Providers
+// Leaflet providers preview: https://leaflet-extras.github.io/leaflet-providers/preview/
 enum EntryMapStyle {
   // Google
   googleNormal,
   googleHybrid,
   googleTerrain,
   // Leaflet
-  // browse providers at https://leaflet-extras.github.io/leaflet-providers/preview/
+  openTopoMap,
   osmHot,
   stamenWatercolor,
 }